divide view methods into mixins
Took 23 minutes
